feat: use full day and month names in date format
- Change weekday from 'short' to 'long' (Monday vs Mon)
- Change month from 'short' to 'long' (October vs Oct)
- Format now displays as 'Monday October 20, 2025'
- Provides more readable and professional date display

+++ src/components/AnimatedNowNextCards.jsx
@@ -49,8 +49,8 @@ const formatDateForDisplay = (dateString) => {
   const date = new Date(year, month - 1, day); // month is 0-indexed in Date constructor
   
   return date.toLocaleDateString('en-US', { 
-    weekday: 'short',
-    month: 'short', 
+    weekday: 'long',
+    month: 'long', 
     day: 'numeric', 
     year: 'numeric' 
   });
